create department modal and other stuff
