--- src/components/CreateRoom/index.js
@@ -1,1 +1,0 @@
-export { default } from './CreateRoom';
