+++ src/screens/ReservationsScreen/Modals/styles.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/config.scss';
+@import '../../../styles/config.scss';
 
 .container {
   @include flexbox($direction: column);
